chore: Moved test cases of BedrockCohereEmbeddingModel into BedrockCohereEmbeddingModelIT
